requirements.txt
streamlit
pillow
qrcode
spacy